@@ -2,7 +2,7 @@
   "dependencies": {
     "@airgap/beacon-dapp": "^4.2.1",
     "@kyvg/vue3-notification": "^3.2.1",
-    "@taquito/rpc": "^19.1.0",
+    "@taquito/rpc": "^19.2.0",
     "vue3-carousel": "0.3.3"
   },
   "description": "bākꜩ - Tezos baking service",
@@ -10,24 +10,25 @@
     "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
     "@nuxt/devtools": "^1.1.5",
     "@nuxt/image": "^1.5.0",
-    "@typescript-eslint/eslint-plugin": "^7.5.0",
-    "@typescript-eslint/parser": "^7.5.0",
-    "eslint": "^8.57.0",
+    "@typescript-eslint/eslint-plugin": "^7.7.0",
+    "@typescript-eslint/parser": "^7.7.0",
+    "eslint": "^9.0.0",
     "eslint-config-ayotte": "^3.2.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-nuxt": "^4.0.0",
     "eslint-plugin-prettier-vue": "^5.0.0",
-    "eslint-plugin-vue": "^9.24.0",
+    "eslint-plugin-vue": "^9.25.0",
     "fs-extra": "^11.2.0",
     "include-media": "^2.0.0",
     "nuxt": "^3.11.2",
     "prettier": "^3.2.5",
-    "sass": "^1.74.1",
-    "sass-loader": "^14.1.1",
+    "sass": "^1.75.0",
+    "sass-loader": "^14.2.1",
     "vite-compatible-readable-stream": "^3.6.1",
     "vue-eslint-parser": "^9.4.2",
-    "vue-tsc": "^2.0.10"
+    "vue-tsc": "^2.0.13"
   },
+  "type": "module",
   "engines": {
     "bun": ">=1.0.21"
   },
